Update README.md
An implementation of the 42 project so_long with all bonus.
It has animations (kind of), an enemy patroling and nice textures with varying trees as walls.

You need minilibx installed to be able to play this game. Don't have it? 42Paris has a nice installation guide. It also provides some maps you can test around with.